rename OCPTasks to OCPSim
